refactor: move toCollections method to ClusterManager
